Renaming projects and changed namespace for Keycloak.IdentityModel
