+++ src/Keycloak.IdentityModel/Extensions/JwtSecurityTokenExtension.cs
@@ -1,0 +1,15 @@
+﻿using Newtonsoft.Json.Linq;
+using System.IdentityModel.Tokens.Jwt;
+using Microsoft.IdentityModel.Tokens;
+
+namespace Keycloak.IdentityModel.Extensions
+{
+    internal static class JwtSecurityTokenExtension
+    {
+        public static JObject GetPayloadJObject(this JwtSecurityToken token)
+        {
+            // TODO: Optimize? Reduce code duplication
+            return JObject.Parse(Base64UrlEncoder.Decode(token.RawPayload));
+        }
+    }
+}
