+++ src/Keycloak.IdentityModel/Global.cs
@@ -1,0 +1,17 @@
+﻿using System.Reflection;
+
+namespace Keycloak.IdentityModel
+{
+    public static class Global
+    {
+        public static string GetVersion()
+        {
+            return Assembly.GetExecutingAssembly().GetName().Version.ToString();
+        }
+
+        public static bool CheckVersion(string version)
+        {
+            return GetVersion() == version;
+        }
+    }
+}
