+++ src/Keycloak.IdentityModel/Keycloak.IdentityModel.nuspec
