+++ src/Keycloak.IdentityModel/Properties/AssemblyInfo.cs
