+++ src/Keycloak.IdentityModel/Utilities/Synchronization/ReaderGuard.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Threading;
+
+namespace Keycloak.IdentityModel.Utilities.Synchronization
+{
+    public class ReaderGuard : IDisposable
+    {
+        private readonly ReaderWriterLockSlim _readerWriterLock;
+
+        public ReaderGuard(ReaderWriterLockSlim readerWriterLock)
+        {
+            _readerWriterLock = readerWriterLock;
+            _readerWriterLock.EnterReadLock();
+        }
+
+        public void Dispose()
+        {
+            _readerWriterLock.ExitReadLock();
+        }
+    }
+}
