--- src/KeycloakIdentityModel/Extensions/DateTimeExtension.cs
@@ -1,25 +1,0 @@
-﻿using System;
-
-namespace KeycloakIdentityModel.Extensions
-{
-    internal static class DateTimeExtension
-    {
-        public static double ToUnixTimestamp(this DateTime dateTime)
-        {
-            return (dateTime - new DateTime(1970, 1, 1).ToLocalTime()).TotalSeconds;
-        }
-
-        public static DateTime ToDateTime(this double unixTimeStamp)
-        {
-            // Unix timestamp is seconds past epoch
-            var dtDateTime = new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc);
-            dtDateTime = dtDateTime.AddSeconds(unixTimeStamp).ToLocalTime();
-            return dtDateTime;
-        }
-
-        public static DateTime ToDateTime(this long unixTimeStamp)
-        {
-            return ToDateTime((double) unixTimeStamp);
-        }
-    }
-}
