--- src/KeycloakIdentityModel/Extensions/JwtSecurityTokenExtension.cs
@@ -1,15 +1,0 @@
-﻿using Newtonsoft.Json.Linq;
-using System.IdentityModel.Tokens.Jwt;
-using Microsoft.IdentityModel.Tokens;
-
-namespace KeycloakIdentityModel.Extensions
-{
-    internal static class JwtSecurityTokenExtension
-    {
-        public static JObject GetPayloadJObject(this JwtSecurityToken token)
-        {
-            // TODO: Optimize? Reduce code duplication
-            return JObject.Parse(Base64UrlEncoder.Decode(token.RawPayload));
-        }
-    }
-}
