--- src/KeycloakIdentityModel/Global.cs
@@ -1,17 +1,0 @@
-﻿using System.Reflection;
-
-namespace KeycloakIdentityModel
-{
-    public static class Global
-    {
-        public static string GetVersion()
-        {
-            return Assembly.GetExecutingAssembly().GetName().Version.ToString();
-        }
-
-        public static bool CheckVersion(string version)
-        {
-            return GetVersion() == version;
-        }
-    }
-}
