+++ src/Owin.Security.Keycloak/Global.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Concurrent;
 using System.Reflection;
-using KeycloakIdentityModel;
-using KeycloakIdentityModel.Utilities.Caching;
+using Keycloak.IdentityModel;
+using Keycloak.IdentityModel.Utilities.Caching;
 
 namespace Owin.Security.Keycloak
 {
