+++ src/Owin.Security.Keycloak/Middleware/KeycloakAuthenticationHandler.cs
@@ -7,8 +7,8 @@ using System.Net;
 using System.Security.Authentication;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using KeycloakIdentityModel;
-using KeycloakIdentityModel.Models.Responses;
+using Keycloak.IdentityModel;
+using Keycloak.IdentityModel.Models.Responses;
 using Microsoft.Owin;
 using Microsoft.Owin.Security;
 using Microsoft.Owin.Security.Infrastructure;
